requirements.txt
streamlit
huggingface_hub
Pillow
numpy